Proposal: move mappers to common package org.pmiops.workbench.mappers
