+++ api/src/test/java/org/pmiops/workbench/api/UserMetricsControllerTest.java
@@ -33,13 +33,13 @@ import org.pmiops.workbench.firecloud.FireCloudService;
 import org.pmiops.workbench.firecloud.model.FirecloudWorkspace;
 import org.pmiops.workbench.firecloud.model.FirecloudWorkspaceResponse;
 import org.pmiops.workbench.google.CloudStorageService;
+import org.pmiops.workbench.mappers.CohortMapper;
+import org.pmiops.workbench.mappers.CohortMapperImpl;
 import org.pmiops.workbench.model.Cohort;
 import org.pmiops.workbench.model.RecentResource;
 import org.pmiops.workbench.model.RecentResourceRequest;
 import org.pmiops.workbench.model.RecentResourceResponse;
 import org.pmiops.workbench.test.FakeClock;
-import org.pmiops.workbench.utils.mappers.CohortMapper;
-import org.pmiops.workbench.utils.mappers.CohortMapperImpl;
 import org.pmiops.workbench.workspaces.WorkspaceService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
